Rename main Markdown block entry point
