+++ client/gutenberg/extensions/presets/jetpack/editor.js
@@ -3,5 +3,5 @@
 /**
  * Internal dependencies
  */
-import 'gutenberg/extensions/markdown/index.jsx';
+import 'gutenberg/extensions/markdown/editor';
 import 'gutenberg/extensions/tiled-gallery/tiled-gallery.jsx';
